refactor : atom 분리
