+++ client/atoms/loginAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const isLoginAtom = atom({
+  key: 'isLogin',
+  default: false,
+});
