+++ client/atoms/reviewAtom.ts
@@ -1,6 +1,5 @@
 import { atom } from 'recoil';
 
-// 채택/후원 관련 atoms 시작
 export const reviewTagsEnumAtom = atom({
   key: 'reviewTagsEnum',
   default: [
@@ -46,46 +45,5 @@ export const reviewRequestAtom = atom({
     articleId: '',
     targetUserName: '',
     dashboardUrl: '',
-  },
-});
-
-// 채택/후원 관련 atoms 끝
-export const isAnswerPostedAtom = atom({
-  key: 'isAnswerPosted',
-  default: false,
-});
-
-export const articleAuthorIdAtom = atom({
-  key: 'articleAuthorId',
-  default: '',
-});
-
-export const isAnswerEditAtom = atom({
-  key: 'isAnswerEdit',
-  default: {
-    isEdit: false,
-    answerId: 0,
-    answerPage: 0,
-    payload: '',
-  },
-});
-
-export const isCommentOpenAtom = atom({
-  key: 'isCommentOpen',
-  default: false,
-});
-
-export const currPagesIndexAtom = atom({
-  key: 'currPagesIndex',
-  default: 0,
-});
-
-export const isArticleEditAtom = atom({
-  key: 'isArticleEdit',
-  default: {
-    isArticleEdit: false,
-    title: '',
-    content: '',
-    articleId: '',
   },
 });
