+++ client/components/common/Header/BtnUser.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { useEffect, useState } from 'react';
 import { useRecoilValue } from 'recoil';
-import { dataHeaderAtom } from '../../../atomsYW';
+import { dataHeaderAtom } from '../../../atoms/userAtom';
 import { changeGradeEmoji } from '../../../libs/changeGradeEmoji';
 
 export const BtnUser = () => {
