+++ client/components/edit-privacy/EditAvatar.tsx
@@ -7,7 +7,9 @@
 import Image from 'next/image';
 import { ChangeEvent, useEffect, useState } from 'react';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
-import { dataHeaderAtom, renderingAtom } from '../../atomsYW';
+import { renderingAtom } from '../../atoms/renderingAtom';
+import { dataHeaderAtom } from '../../atoms/userAtom';
+
 import { client } from '../../libs/client';
 import { uploadImg } from '../../libs/uploadS3';
