+++ client/libs/useCheckIsLogin.ts
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import { useRecoilValue } from 'recoil';
-import { isLoginAtom } from '../atomsYW';
+import { isLoginAtom } from '../atoms/loginAtom';
 
 export const useCheckClickIsLogin = () => {
   const isLogin = useRecoilValue(isLoginAtom);
