+++ client/pages/post/index.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { toast } from 'react-toastify';
 import { useRecoilValue } from 'recoil';
-import { isLoginAtom } from '../../atomsYW';
+import { isLoginAtom } from '../../atoms/loginAtom';
 import { Editor } from '../../components/common/QuillEditor/Editor';
 import { Seo } from '../../components/common/Seo';
