+++ client/pages/review/support-morak.tsx
@@ -12,7 +12,7 @@ import {
   reviewRequestAtom,
   reviewTagsAtom,
   reviewPointAtom,
-} from '../../atomsHJ';
+} from '../../atoms/reviewAtom';
 import { ProgressBar } from '../../components/review/ProgressBar';
 import { BtnBackArticle } from '../../components/review/BtnBackArticle';
 import { client } from '../../libs/client';
